+++ dimGray/src/main/java/urlshortener2015/dimGray/Application.java
@@ -1,4 +1,4 @@
-package urlshortener2015.demo;
+package urlshortener2015.dimGray;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
