+++ settings.gradle
@@ -3,7 +3,7 @@ include ":demo"
 include ":alizarin-crimson"
 include ":bangladesh-green"
 include ":candy-pink"
-include ":dim-gray"
+include ":dimGray"
 include ":eerie-black"
 include ":fuzzy-wuzzy"
 include ":golden-brown"
